+++ pymince/logging.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+
 import contextlib
 import json
 import logging