+++ tests/test_logging/test_timed_block.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+
 import unittest.mock
 
 import pymince.logging
@@ -9,5 +10,4 @@ def test_timed_block():
     with pymince.logging.timed_block("Testing", logger=logger):
         pass
 
-    assert logger.info.called
-    assert logger.debug.called
+    assert all((logger.info.called, logger.debug.called))