+++ tests/test_std/test_bind_json_std.py
@@ -34,4 +34,4 @@ def test_bind_json_std_without_output(capsys, stub_stdin_payload):
     pymince.std.bind_json_std()(function)()
     captured = capsys.readouterr()
     function.assert_called_once_with(data=stub_stdin_payload)
-    assert captured.out == ""
+    assert not captured.out